+++ src/daf/message/text_based.py
@@ -41,6 +41,13 @@ sql.register_type("MessageMODE", "clear-send")
 @sql.register_type("MessageTYPE")
 class TextMESSAGE(BaseChannelMessage):
     """
+    .. versionchanged:: v4.2.0
+
+        - Removed deprecated ``start_in``,
+        ``start_period`` and ``end_period`` parameters.
+        - Removed support for ``data``'s datatypes that are not
+        :class:`daf.messagedata.BaseTextData`.
+
     This class is used for creating objects that represent messages which will be sent to Discord's TEXT CHANNELS.
 
     Parameters
@@ -120,28 +127,10 @@ class TextMESSAGE(BaseChannelMessage):
         if not isinstance(data, BaseTextData):
             trace(
                 f"Using data types other than {[x.__name__ for x in BaseTextData.__subclasses__()]}, "
-                "is deprecated on TextMESSAGE's data parameter! Planned for removal in 4.2.0",
-                TraceLEVELS.DEPRECATED
+                "is deprecated on TextMESSAGE's data parameter! And has been removed",
+                TraceLEVELS.ERROR,
+                exception_cls=TypeError
             )
-            # Transform to new data type            
-            if isinstance(data, _FunctionBaseCLASS):
-                data = _DeprecatedDynamic(data.fnc, *data.args, **data.kwargs)
-            elif data is not None:
-                if isinstance(data, (str, discord.Embed, FILE)):
-                    data = [data]
-
-                content = None
-                embed = None
-                files = []
-                for item in data:
-                    if isinstance(item, str):
-                        content = item
-                    elif isinstance(item, discord.Embed):
-                        embed = item
-                    else:
-                        files.append(item)
-
-                data = TextMessageData(content, embed, files)
 
         super().__init__(start_period, end_period, data, channels, start_in, remove_after, period)
 
@@ -491,6 +480,13 @@ class DirectMESSAGE(BaseMESSAGE):
     """
     This class is used for creating objects that represent messages which will be sent to user's private messages.
 
+    .. versionchanged:: v4.2.0
+
+        - Removed deprecated ``start_in``,
+        ``start_period`` and ``end_period`` parameters.
+        - Removed support for ``data``'s datatypes that are not
+        :class:`daf.messagedata.BaseTextData`.
+
     .. deprecated:: v2.1
 
         - start_period, end_period - Using int values, use ``timedelta`` object instead.
@@ -542,28 +538,10 @@ class DirectMESSAGE(BaseMESSAGE):
         if not isinstance(data, BaseTextData):
             trace(
                 f"Using data types other than {[x.__name__ for x in BaseTextData.__subclasses__()]}, "
-                "is deprecated on DirectMESSAGE's data parameter! Planned for removal in 4.2.0",
-                TraceLEVELS.DEPRECATED
+                "is deprecated on TextMESSAGE's data parameter! And has been removed",
+                TraceLEVELS.ERROR,
+                exception_cls=TypeError
             )
-            # Transform to new data type            
-            if isinstance(data, _FunctionBaseCLASS):
-                data = _DeprecatedDynamic(data.fnc, *data.args, *data.kwargs)
-            elif data is not None:
-                if isinstance(data, (str, discord.Embed, FILE)):
-                    data = [data]
-
-                content = None
-                embed = None
-                files = []
-                for item in data:
-                    if isinstance(item, str):
-                        content = item
-                    elif isinstance(item, discord.Embed):
-                        embed = item
-                    else:
-                        files.append(item)
-
-                data = TextMessageData(content, embed, files)
 
         super().__init__(start_period, end_period, data, start_in, remove_after, period)
         self.mode = mode
@@ -578,7 +556,6 @@ class DirectMESSAGE(BaseMESSAGE):
             self._remove_after -= 1
             if not self._remove_after:
                 self._event_ctrl.emit(EventID._trigger_message_remove, self.parent, self)
-
 
     def generate_log_context(self,
                              success_context: Dict[str, Union[bool, Optional[Exception]]],
@@ -765,15 +742,12 @@ class DirectMESSAGE(BaseMESSAGE):
 
     async def _on_update(self, _, _init_options: Optional[dict], **kwargs):
         await self._close()
-        if "start_in" not in kwargs:
-            # This parameter does not appear as attribute, manual setting necessary
-            kwargs["start_in"] = None
 
-        if "start_period" not in kwargs:  # DEPRECATED, TODO: REMOVE IN FUTURE
-            kwargs["start_period"] = None
-
-        if "end_period" not in kwargs:  # DEPRECATED, TODO: REMOVE IN FUTURE
-            kwargs["end_period"] = None
+         # DEPRECATED, TODO: REMOVE IN FUTURE
+        # These parameters do not appear as attribute, manual setting necessary.
+        kwargs["start_in"] = None
+        kwargs["start_period"] = None
+        kwargs["end_period"] = None
 
         if "data" not in kwargs:
             kwargs["data"] = self._data